Add files via upload
This is a database holding a table called "controller" with two columns called "a" and "b". a is in charge of handling the addresses, for prototyping a has addresses 1 through 8. b is in charge of handling the IP addresses, all IP addresses are the same for Doug's home network allocation of the MKR  MAKE SURE TO UPDATE THE SELECTED IP ADDRESS TO BE A WORKABLE ONE FOR YOU.